REFACTOR. Use single source of 'files'
Removes references to 'previewItems' and 
'uploadItems' and uses a single 'files' array
on state.

Files being uploaded are prepended to the 'files'
array so they're always rendered above existing
uploaded files.

This 'order' allows us to remove and delete an XHR
of a an item by it's index in the array.