Manually check for content in the richTextArea before propagating data.
This fixes #82. The issue is that empty content blocks are representable
in the draft-js AST we create, so when it appears there’s no content
the editor actually has an empty unstyled block. We now check if there’s
any text in the editor using `hasText` and only persist the AST data up
if there is some text. At the moment this works because every object
within draft-js has to have text (even atomic blocks). If that changes
we might need to adjust things to do a more intelligent check.