Debounce the onChange by a millisecond to avoid timing issue
There appears to be a timing issue in the editor that results in multiple but
slightly out of sync changes being called when:

- A formalist atomic block is rendered that has a state change in its
  component, e.g., atomic/index.js -> setState({isSelected})
- User presses enter to create a new block
- User presses up

The problem goes away when the exported editor state isn’t passed back to the
formalist AST, which makes it _seem_ like it’s tied to rendering but the
`editorState` object doesn’t get re-interpreted and so that seems odd. At a
guess, it’s a timing issue releated to setState being async and so when the AST
is updated and the component is rerendered it renders twice in the same moment:
one with the new state and the with the old state.