Fix bug where atomic blocks would duplicate
Special handling for when the enter key is pressed while an atomic block is
selected. This appears to be a draft-js bug. When this situation occurs the
atomic block duplicated (though its entity does not). There's no way to do this
so we just have to *immediately* reverse its function