Trigger a focus event on the dropzone when its interacted with
This fixes an issue with upload fields in embedded forms. Because
clicking or dragging onto the drop zone wouldn’t trigger a focus
the wrapping draft-js instance wouldn’t be set to readOnly and so would
see some of the interaction as “text changes” and destroy the formalist
entity definition somehow.