Send Content-Type through to presign requests.
This means that presign requests can enforce the Content-Type as they wish, or pass it through as a parameter to S3 which means the object has that same Content-Type, rather than the generic application/octet-stream.

I'm not sure if there's a better way to clone the presign params, but a shallow clone is probably good enough. Certainly is fine for my existing usage.